RembgBackend/requirements.txt
flask
flask-cors
rembg
Pillow
opencv-python
numpy
pytesseract
scikit-learn




